fix: Event update when it is dragged and dropped now persists on the database!
